requirements.txt
pyqt5
python-docx
python-pptx
openpyxl
xlrd
xlwt
chardet
pywin32 